Searching large scroll buffers impairs standard laptop
When searching a large scroll buffer, and each letter added to the
search term is searched for impairs the speed and stalls the zellij
terminal. Therefore, removing this automatic search will improve
workflow when searching.

Change:
* Removing automatic search on each added letter to the search term.
* This now requires people to hit the `Enter key` and then press search
up `p` or down `n` before the search term appears in the pane view.